refactor Welcome2Screen
